Resource sharing, closes #101
This change simplifies Resources. Now they can be declared as members of
the test class and be used in `withResource` decorators. It's also
possible to derive new resources from existing ones.

To make test discovery cheaper, the test class is no longer instantiated
in that step, rendering the assertion name useless. Consequently the
assertion name method has been removed for now. It may have to be added
back for support of parametrized tests, though.