unsafeInstanceOf Quality, closes #155
This provides a Quality similar to `istanceOf` but also taking
`Quality`s of subclasses of the given class. This is unsafe because
there is no guarantee that the testee is actually of that subtype, but
it may be required if the testee is a generic class.

It solves issues like described in
https://github.com/hamcrest/JavaHamcrest/issues/388