implement assessment pass description, closes 202 (#211)
With this change an `Assessment` always provides a description, also in the successful case. This is useful for `Quality`s like `not`, which can give a better description now. Before, `not` could only return the actual value and the expectation of the delegate `Quality`. Now it can return a description of the actual `Assessment`, which often contains more useful information.

To support development of pass descriptions, the `passes` `Quality` now has a variant that takes a single value and the expected description.

Also, all `Quality`s have been updated to return proper pass descriptions, although some need further improvements in future updates.

Some deprecated classes and constructors have been removed.